[Simplification] Remove conventions and source sets in favour of simple extensions
This replaces the "Conventions" and "custom source sets" in favour of
using a simple task + `extension` on any `SourceSet` that are defined.
This *greatly* simplifies the code and matches all tests still. Further,
this properly adds support for incremental changes in templates as they
were not correctly implemented before.

This was extracted from #17.