[Simplify] Remove custom source set implementation in favour of using extension directly attached to `sourceSet`
This removes the custom `AssertJSourceSet` which was not adding value with the new functionality in modern gradle that generates methods to configure extensions for those that they are attached to. For example, the `assertJ {}` closure is handled by the `extension` rather than adding a new `SourceSet` type.

This was extracted from #31