[CI] Add Github publishing (#49)
This adds release support to Github. When we create a release `tag` this will publish a new version of the plugin. For now, this does not publish `SNAPSHOT` builds as there's currently lots of work happening here.

This does require us to add the correct environment variables to this repo.